# services/core-api/continuum_api_spec.yaml
openapi: 3.0.3
...
info:
  title: Continuum CRM API
  version: '1.0.0'
  description: |
    An AI-native CRM that captures customer interactions automatically and
    proposes the resulting record changes for human confirmation.

    ## Authentication

    All endpoints except signup, login, provider OAuth callbacks and provider
    webhooks require a bearer token:

        Authorization: Bearer <token>

    Two token types are accepted:

    - **Session JWT** — issued by `/auth/login`, used by the web application.
    - **API token** — prefixed `ctm_`, created in Settings. Subject to the rate
      limit below, and revocable with effect on the very next request.

    ## Rate limiting

    API tokens are limited to 300 requests per minute (FR-API-01). Every
    response carries `X-RateLimit-Limit`, `X-RateLimit-Remaining` and
    `X-RateLimit-Reset`. Requests below the limit are never throttled.

    ## Provenance

    Records that the agent can infer carry a `source` field
    (`HUMAN` | `AGENT_INFERRED` | `IMPORTED` | `ENRICHED`). Deals additionally
    carry `stageSource`, recording whether the current stage was set by a
    person or by a confirmed agent proposal. Consuming clients should surface
    this distinction rather than flattening it.

  contact:
    name: Continuum engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.continuum.example/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Query' }
        - { $ref: '#/components/parameters/Tag' }
      responses:
        '200':
          description: Contacts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Contact' }
    post:
      tags: [Contacts]
      summary: Create a contact
      description: |
        Contacts created here are `source: HUMAN`. A contact whose email domain
        matches a known company is linked automatically (FR-DATA-03).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
  /contacts/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Contacts]
      summary: Get a contact
      responses:
        '200':
          description: Contact
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Contacts]
      summary: Update a contact
      description: Any human edit reclaims the field — `source` becomes `HUMAN`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
    delete:
      tags: [Contacts]
      summary: Delete a contact
      responses:
        '204': { description: Deleted }
  /contacts/{id}/activities:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Contacts, Activities]
      summary: Signal Thread for a contact
      description: The unified chronological timeline (FR-AC-05).
      responses:
        '200':
          description: Activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Activity' }
...
  /deals:
    get:
      tags: [Deals]
      summary: List deals
      description: |
        Scoped by role: members see their own, managers their team's, admins the
        whole workspace (FR-PERM-03).
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Query' }
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, WON, LOST] }
        - name: pipelineId
          in: query
          schema: { type: string, format: uuid }
        - name: stageId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Deals
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Deal' }
    post:
      tags: [Deals]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                pipelineId: { type: string, format: uuid }
                stageId: { type: string, format: uuid }
                valueCents: { type: integer, nullable: true }
                companyId: { type: string, format: uuid, nullable: true }
                contactIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deal' }
  /deals/{id}/stage:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Deals]
      summary: Move a deal to a stage
      description: |
        Separate from `PATCH /deals/{id}` so `stageSource` is unambiguously set
        to `HUMAN` and cannot be changed by accident (FR-PIPE-01). Landing on a
        won stage sets status to `WON` in the same action, making the deal
        eligible for project conversion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stageId]
              properties:
                stageId: { type: string, format: uuid }
      responses:
        '200':
          description: Moved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deal' }
  /deals/{id}/activities:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Deals, Activities]
      summary: Signal Thread for a deal
      responses:
        '200':
          description: Activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Activity' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session JWT, or a `ctm_`-prefixed API token.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Query:
      name: q
      in: query
      schema: { type: string }
    Tag:
      name: tag
      in: query
      schema: { type: string }
  responses:
...
    NotFound:
      description: No such record
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable machine-readable code — branch on this, not on the message.
            message: { type: string }
            details: {}
    Paginated:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    RecordSource:
      type: string
      description: |
        Where a value came from. Clients should surface this distinction — an
        agent-inferred value must never be presented as indistinguishable from
        one a person typed (FR-DATA-02).
      enum: [HUMAN, AGENT_INFERRED, IMPORTED, ENRICHED]
...
    Contact:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        fullName: { type: string }
        title: { type: string, nullable: true }
        companyId: { type: string, format: uuid, nullable: true }
        source: { $ref: '#/components/schemas/RecordSource' }
        excluded:
          type: boolean
          description: Capture is switched off for this sender (FR-AC-08).
    ContactInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        title: { type: string, nullable: true }
        companyId: { type: string, format: uuid, nullable: true }
        tags: { type: array, items: { type: string } }
    Deal:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        pipelineId: { type: string, format: uuid }
        stageId: { type: string, format: uuid }
        stageSource:
          allOf: [{ $ref: '#/components/schemas/RecordSource' }]
          description: |
            Stored separately from `stageId` so the UI can always answer
            "human or agent?" without inspecting the audit log.
        status: { type: string, enum: [OPEN, WON, LOST] }
        valueCents: { type: integer, nullable: true }
        lastActivityAt: { type: string, format: date-time, nullable: true }
        stallingSince:
          type: string
          format: date-time
          nullable: true
          description: Non-null while past the workspace inactivity threshold (FR-AC-06).
        convertedProjectId: { type: string, format: uuid, nullable: true }
        needsReview:
          type: boolean
          description: |
            Set when a converted deal was reopened — the relationship is flagged
            for a human rather than silently reverted.
    Activity:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [EMAIL, CALL, MEETING, NOTE] }
        direction: { type: string, enum: [INBOUND, OUTBOUND, INTERNAL] }
        subject: { type: string, nullable: true }
        body:
          type: string
          nullable: true
          description: |
            An excerpt, not the full message. Continuum stores enough to audit
            the summary against; the complete text stays in the user's mailbox
            (data minimization, §30).
        occurredAt: { type: string, format: date-time }
        aiSummary:
          type: string
          nullable: true
          description: Null whenever summarization has not succeeded — never fabricated.
        sentiment:
          type: string
          nullable: true
          enum: [POSITIVE, NEUTRAL, NEGATIVE]
        nextStep:
          type: string
          nullable: true
          description: |
            Explicitly nullable. "No next step" is representable, so the model is
            never forced to invent one (FR-AC-03).
        summaryStatus: { type: string, enum: [PENDING, DONE, FAILED, SKIPPED] }
        callOutcome:
          type: string
          nullable: true
          enum: [CONNECTED, ATTEMPTED_NO_CONNECT, VOICEMAIL, FAILED]
        source: { $ref: '#/components/schemas/RecordSource' }
...